Extract help, init, and install commands from cli/index.ts
