+++ cli/commands/index.ts
@@ -1,0 +1,3 @@
+export { helpCommand } from "df/cli/commands/help_command";
+export { initCommand, icebergOption } from "df/cli/commands/init_command";
+export { installCommand } from "df/cli/commands/install_command";
